Fix TypeScript generator generating invalid keys
When one has a dictionary that has keys of non-string type that are
also not enums, they should be generated as string keys on the
TypeScript side.

The correct type is resolved (`string`) but an interface prefix is added
for some reason, which breaks the generated TypeScript since there is
no type called `Istring`.

This bug seems to have been introduced with commit
543724234ea7fc97e1d81ce7cc9fd2ca76c32fa8
from pull request #755.

This commit should fix issue #898.