Support DataContract private fields and properties
In f7f530c54f7b8970104fcba21156d02518885057 a fix was put in to
fix a crash when a property or field did not have MemberInfo.
Unfortunately, this also filtered out private fields and properties
since the check only searched the public fields.  This commit
re-allows private properties and fields by searching the entire
type when looking for a MemberInfo for a property.