Upgrade net45 tests to net452 to satisfy xunit
The version of xunit currently used does not support .Net 4.5 anymore, and some tests are not running because of that.
All the net45 tests are upgraded to net452 to satisfy xunit.
I also refactored an obsolete test using CodeDomProvider instead of Roslyn in InheritanceSerializarionTests.cs. The old compiler can't compile read-only properties in the generated code.