Use DisplayAttribute methods when getting name and description
`DisplayAttribute` has `GetName` and `GetDescription` methods that will return the `Name` and `Description` property values respectively if no `ResourceType` has been specified on the attribute, or will look up resource values if a `ResourceType` has been specified.

This commit updates schema generation to use the above methods when a `DisplayAttribute` is used.